requirements.pred.txt
networkx==2.5.1
pydot==1.4.2
pyyaml==6
pygraphviz==1.6
funcy==1.16
scikit-learn==0.24.2
click==8.0.1
